admin authentication

+++ src/client/components/Reviews.jsx
@@ -74,22 +74,30 @@ export default function Reviews({token}){
    
   
   // delete review by id function 
-  async function deleteReview (reviewId) {
-  
-    try {
-      const response = await fetch(`http://localhost:3000/api/reviews/${reviewId}`, {
-        method: "DELETE",
-        headers: {
-          'Content-Type': 'application/json',
-          // 'Authorization': `Bearer ${tokenString}`
-        }
-    });
-    const resultDelete = await response.json();
-    // if (resultDelete) {window.location.reload()}
-    alert("Review successfully deleted.");
-  } catch (error) {
-    console.error(error);
-  }}; 
+ async function deleteReview(reviewId, isAdmin) {
+      try{
+        if (isAdmin === 'true') {
+          const response = await fetch(`http://localhost:3000/api/reviews/${reviewId}`, {
+            method: "DELETE",
+            headers: {
+              'Content-Type': 'application/json',
+              // 'Authorization': `Bearer ${tokenString}`
+            }
+        });
+
+      const resultDelete = await response.json();
+      if (response.status === 200) {
+        alert("Review successfully deleted!")  
+      } else {
+        alert("Failed to delete review.")
+      } 
+      } else {
+        alert("You are not authorized to delete this review")
+      }
+       } catch (error){
+        console.error(error);
+       }
+ }
 
 // const reviewBoxStyle = {
 //   border: '1px solid #ccc',
